Some refactoring

+++ Game/CollisionEvent.h
@@ -8,7 +8,7 @@ struct CollisionEvent {
 	Entity entity = { -1 };
 	Entity other = { -1 };
 	Vector3D collisionDirection = { 0, 0, 0 };
-	double time = 1.01;
+	float time = 1.01;
 };
 
 using CollisionEventMap = std::unordered_map<int, DynamicArray<CollisionEvent>>;
