+++ Game/Timer.h
@@ -5,5 +5,5 @@ class Timer {
 	std::chrono::time_point<std::chrono::high_resolution_clock> startTime;
 public:
 	void start();
-	double stop();
+	float stop();
 };
